@@ -1,31 +1,14 @@
-package lib
+package snowflake_client
 
 import (
-	"bytes"
 	"fmt"
-	"io/ioutil"
 	"net"
-	"net/http"
 	"testing"
+	"time"
 
-	"git.torproject.org/pluggable-transports/snowflake.git/common/util"
 	. "github.com/smartystreets/goconvey/convey"
+	"gitlab.torproject.org/tpo/anti-censorship/pluggable-transports/snowflake/v2/common/event"
 )
-
-type MockTransport struct {
-	statusOverride int
-	body           []byte
-}
-
-// Just returns a response with fake SDP answer.
-func (m *MockTransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	s := ioutil.NopCloser(bytes.NewReader(m.body))
-	r := &http.Response{
-		StatusCode: m.statusOverride,
-		Body:       s,
-	}
-	return r, nil
-}
 
 type FakeDialer struct {
 	max int
@@ -33,7 +16,7 @@ type FakeDialer struct {
 
 func (w FakeDialer) Catch() (*WebRTCPeer, error) {
 	fmt.Println("Caught a dummy snowflake.")
-	return &WebRTCPeer{}, nil
+	return &WebRTCPeer{closed: make(chan struct{})}, nil
 }
 
 func (w FakeDialer) GetMax() int {
@@ -50,12 +33,6 @@ func (f FakeSocksConn) Reject() error {
 	return nil
 }
 func (f FakeSocksConn) Grant(addr *net.TCPAddr) error { return nil }
-
-type FakePeers struct{ toRelease *WebRTCPeer }
-
-func (f FakePeers) Collect() (*WebRTCPeer, error) { return &WebRTCPeer{}, nil }
-func (f FakePeers) Pop() *WebRTCPeer              { return nil }
-func (f FakePeers) Melted() <-chan struct{}       { return nil }
 
 func TestSnowflakeClient(t *testing.T) {
 
@@ -97,7 +74,7 @@ func TestSnowflakeClient(t *testing.T) {
 			So(err, ShouldNotBeNil)
 			So(p.Count(), ShouldEqual, c)
 
-			// But popping and closing allows it to continue.
+			// But popping allows it to continue.
 			s := p.Pop()
 			s.Close()
 			So(s, ShouldNotBeNil)
@@ -127,7 +104,7 @@ func TestSnowflakeClient(t *testing.T) {
 			cnt := 5
 			p, _ := NewPeers(FakeDialer{max: cnt})
 			for i := 0; i < cnt; i++ {
-				p.activePeers.PushBack(&WebRTCPeer{})
+				p.activePeers.PushBack(&WebRTCPeer{closed: make(chan struct{})})
 			}
 			So(p.Count(), ShouldEqual, cnt)
 			p.End()
@@ -154,31 +131,36 @@ func TestSnowflakeClient(t *testing.T) {
 			So(r, ShouldEqual, wc4)
 		})
 
-	})
+		Convey("Terminate Connect() loop", func() {
+			p, _ := NewPeers(FakeDialer{max: 4})
+			go func() {
+				for {
+					p.Collect()
+					select {
+					case <-p.Melted():
+						return
+					default:
+					}
+				}
+			}()
+			<-time.After(10 * time.Second)
 
-	Convey("Snowflake", t, func() {
-
-		SkipConvey("Handler Grants correctly", func() {
-			socks := &FakeSocksConn{}
-			broker := &BrokerChannel{Host: "test"}
-			d := NewWebRTCDialer(broker, nil, 1)
-
-			So(socks.rejected, ShouldEqual, false)
-			Handler(socks, d)
-			So(socks.rejected, ShouldEqual, true)
+			p.End()
+			<-p.Melted()
+			So(p.Count(), ShouldEqual, 0)
 		})
+
 	})
 
 	Convey("Dialers", t, func() {
 		Convey("Can construct WebRTCDialer.", func() {
-			broker := &BrokerChannel{Host: "test"}
+			broker := &BrokerChannel{}
 			d := NewWebRTCDialer(broker, nil, 1)
 			So(d, ShouldNotBeNil)
 			So(d.BrokerChannel, ShouldNotBeNil)
-			So(d.BrokerChannel.Host, ShouldEqual, "test")
 		})
 		SkipConvey("WebRTCDialer can Catch a snowflake.", func() {
-			broker := &BrokerChannel{Host: "test"}
+			broker := &BrokerChannel{}
 			d := NewWebRTCDialer(broker, nil, 1)
 			conn, err := d.Catch()
 			So(conn, ShouldBeNil)
@@ -186,84 +168,54 @@ func TestSnowflakeClient(t *testing.T) {
 		})
 	})
 
-	Convey("Rendezvous", t, func() {
-		transport := &MockTransport{
-			http.StatusOK,
-			[]byte(`{"type":"answer","sdp":"fake"}`),
-		}
-		fakeOffer, err := util.DeserializeSessionDescription(`{"type":"offer","sdp":"test"}`)
-		if err != nil {
-			panic(err)
-		}
+}
 
-		Convey("Construct BrokerChannel with no front domain", func() {
-			b, err := NewBrokerChannel("test.broker", "", transport, false)
-			So(b.url, ShouldNotBeNil)
-			So(err, ShouldBeNil)
-			So(b.url.Path, ShouldResemble, "test.broker")
-			So(b.transport, ShouldNotBeNil)
-		})
-
-		Convey("Construct BrokerChannel *with* front domain", func() {
-			b, err := NewBrokerChannel("test.broker", "front", transport, false)
-			So(b.url, ShouldNotBeNil)
-			So(err, ShouldBeNil)
-			So(b.url.Path, ShouldResemble, "test.broker")
-			So(b.url.Host, ShouldResemble, "front")
-			So(b.transport, ShouldNotBeNil)
-		})
-
-		Convey("BrokerChannel.Negotiate responds with answer", func() {
-			b, err := NewBrokerChannel("test.broker", "", transport, false)
-			So(err, ShouldBeNil)
-			answer, err := b.Negotiate(fakeOffer)
-			So(err, ShouldBeNil)
-			So(answer, ShouldNotBeNil)
-			So(answer.SDP, ShouldResemble, "fake")
-		})
-
-		Convey("BrokerChannel.Negotiate fails with 503", func() {
-			b, err := NewBrokerChannel("test.broker", "",
-				&MockTransport{http.StatusServiceUnavailable, []byte("\n")},
-				false)
-			So(err, ShouldBeNil)
-			answer, err := b.Negotiate(fakeOffer)
-			So(err, ShouldNotBeNil)
-			So(answer, ShouldBeNil)
-			So(err.Error(), ShouldResemble, BrokerError503)
-		})
-
-		Convey("BrokerChannel.Negotiate fails with 400", func() {
-			b, err := NewBrokerChannel("test.broker", "",
-				&MockTransport{http.StatusBadRequest, []byte("\n")},
-				false)
-			So(err, ShouldBeNil)
-			answer, err := b.Negotiate(fakeOffer)
-			So(err, ShouldNotBeNil)
-			So(answer, ShouldBeNil)
-			So(err.Error(), ShouldResemble, BrokerError400)
-		})
-
-		Convey("BrokerChannel.Negotiate fails with large read", func() {
-			b, err := NewBrokerChannel("test.broker", "",
-				&MockTransport{http.StatusOK, make([]byte, 100001, 100001)},
-				false)
-			So(err, ShouldBeNil)
-			answer, err := b.Negotiate(fakeOffer)
-			So(err, ShouldNotBeNil)
-			So(answer, ShouldBeNil)
-			So(err.Error(), ShouldResemble, "unexpected EOF")
-		})
-
-		Convey("BrokerChannel.Negotiate fails with unexpected error", func() {
-			b, err := NewBrokerChannel("test.broker", "",
-				&MockTransport{123, []byte("")}, false)
-			So(err, ShouldBeNil)
-			answer, err := b.Negotiate(fakeOffer)
-			So(err, ShouldNotBeNil)
-			So(answer, ShouldBeNil)
-			So(err.Error(), ShouldResemble, BrokerErrorUnexpected)
+func TestWebRTCPeer(t *testing.T) {
+	Convey("WebRTCPeer", t, func(c C) {
+		p := &WebRTCPeer{closed: make(chan struct{}),
+			eventsLogger: event.NewSnowflakeEventDispatcher()}
+		Convey("checks for staleness", func() {
+			go p.checkForStaleness(time.Second)
+			<-time.After(2 * time.Second)
+			So(p.Closed(), ShouldEqual, true)
 		})
 	})
+}
 
+func TestICEServerParser(t *testing.T) {
+	Convey("Test parsing of ICE servers", t, func() {
+		for _, test := range []struct {
+			input  []string
+			urls   [][]string
+			length int
+		}{
+			{
+				[]string{"stun:stun.l.google.com:19302", "stun:stun.ekiga.net"},
+				[][]string{[]string{"stun:stun.l.google.com:19302"}, []string{"stun:stun.ekiga.net:3478"}},
+				2,
+			},
+			{
+				[]string{"stun:stun1.l.google.com:19302", "stun.ekiga.net", "stun:stun.example.com:1234/path?query",
+					"https://example.com", "turn:relay.metered.ca:80?transport=udp"},
+				[][]string{[]string{"stun:stun1.l.google.com:19302"}},
+				1,
+			},
+		} {
+			servers := parseIceServers(test.input)
+
+			if test.urls == nil {
+				So(servers, ShouldBeNil)
+			} else {
+				So(servers, ShouldNotBeNil)
+			}
+
+			So(len(servers), ShouldEqual, test.length)
+
+			for _, server := range servers {
+				So(test.urls, ShouldContain, server.URLs)
+			}
+
+		}
+
+	})
 }
