@@ -1,4 +1,4 @@
-package lib
+package snowflake_client
 
 import (
 	"bufio"
@@ -7,15 +7,15 @@ import (
 	"net"
 	"time"
 
-	"git.torproject.org/pluggable-transports/snowflake.git/common/encapsulation"
+	"gitlab.torproject.org/tpo/anti-censorship/pluggable-transports/snowflake/v2/common/encapsulation"
 )
 
 var errNotImplemented = errors.New("not implemented")
 
-// EncapsulationPacketConn implements the net.PacketConn interface over an
+// encapsulationPacketConn implements the net.PacketConn interface over an
 // io.ReadWriteCloser stream, using the encapsulation package to represent
 // packets in a stream.
-type EncapsulationPacketConn struct {
+type encapsulationPacketConn struct {
 	io.ReadWriteCloser
 	localAddr  net.Addr
 	remoteAddr net.Addr
@@ -23,11 +23,11 @@ type EncapsulationPacketConn struct {
 }
 
 // NewEncapsulationPacketConn makes
-func NewEncapsulationPacketConn(
+func newEncapsulationPacketConn(
 	localAddr, remoteAddr net.Addr,
 	conn io.ReadWriteCloser,
-) *EncapsulationPacketConn {
-	return &EncapsulationPacketConn{
+) *encapsulationPacketConn {
+	return &encapsulationPacketConn{
 		ReadWriteCloser: conn,
 		localAddr:       localAddr,
 		remoteAddr:      remoteAddr,
@@ -36,7 +36,7 @@ func NewEncapsulationPacketConn(
 }
 
 // ReadFrom reads an encapsulated packet from the stream.
-func (c *EncapsulationPacketConn) ReadFrom(p []byte) (int, net.Addr, error) {
+func (c *encapsulationPacketConn) ReadFrom(p []byte) (int, net.Addr, error) {
 	data, err := encapsulation.ReadData(c.ReadWriteCloser)
 	if err != nil {
 		return 0, c.remoteAddr, err
@@ -45,7 +45,7 @@ func (c *EncapsulationPacketConn) ReadFrom(p []byte) (int, net.Addr, error) {
 }
 
 // WriteTo writes an encapsulated packet to the stream.
-func (c *EncapsulationPacketConn) WriteTo(p []byte, addr net.Addr) (int, error) {
+func (c *encapsulationPacketConn) WriteTo(p []byte, addr net.Addr) (int, error) {
 	// addr is ignored.
 	_, err := encapsulation.WriteData(c.bw, p)
 	if err == nil {
@@ -59,10 +59,10 @@ func (c *EncapsulationPacketConn) WriteTo(p []byte, addr net.Addr) (int, error) 
 
 // LocalAddr returns the localAddr value that was passed to
 // NewEncapsulationPacketConn.
-func (c *EncapsulationPacketConn) LocalAddr() net.Addr {
+func (c *encapsulationPacketConn) LocalAddr() net.Addr {
 	return c.localAddr
 }
 
-func (c *EncapsulationPacketConn) SetDeadline(t time.Time) error      { return errNotImplemented }
-func (c *EncapsulationPacketConn) SetReadDeadline(t time.Time) error  { return errNotImplemented }
-func (c *EncapsulationPacketConn) SetWriteDeadline(t time.Time) error { return errNotImplemented }
+func (c *encapsulationPacketConn) SetDeadline(t time.Time) error      { return errNotImplemented }
+func (c *encapsulationPacketConn) SetReadDeadline(t time.Time) error  { return errNotImplemented }
+func (c *encapsulationPacketConn) SetWriteDeadline(t time.Time) error { return errNotImplemented }
