@@ -11,7 +11,7 @@ import "errors"
 var Token = [8]byte{0x12, 0x93, 0x60, 0x5d, 0x27, 0x81, 0x75, 0xf5}
 
 // The size of receive and send queues.
-const queueSize = 32
+const queueSize = 512
 
 var errClosedPacketConn = errors.New("operation on closed connection")
 var errNotImplemented = errors.New("not implemented")
