@@ -42,6 +42,7 @@ func (conn *Conn) SetDeadline(t time.Time) error {
 }
 
 func readLoop(w io.Writer, ws *websocket.Conn) error {
+	var buf [2048]byte
 	for {
 		messageType, r, err := ws.NextReader()
 		if err != nil {
@@ -50,7 +51,7 @@ func readLoop(w io.Writer, ws *websocket.Conn) error {
 		if messageType != websocket.BinaryMessage && messageType != websocket.TextMessage {
 			continue
 		}
-		_, err = io.Copy(w, r)
+		_, err = io.CopyBuffer(w, r, buf[:])
 		if err != nil {
 			return err
 		}
@@ -58,22 +59,13 @@ func readLoop(w io.Writer, ws *websocket.Conn) error {
 }
 
 func writeLoop(ws *websocket.Conn, r io.Reader) error {
+	var buf [2048]byte
 	for {
-		var buf [2048]byte
 		n, err := r.Read(buf[:])
 		if err != nil {
 			return err
 		}
-		data := buf[:n]
-		w, err := ws.NextWriter(websocket.BinaryMessage)
-		if err != nil {
-			return err
-		}
-		n, err = w.Write(data)
-		if err != nil {
-			return err
-		}
-		err = w.Close()
+		err = ws.WriteMessage(websocket.BinaryMessage, buf[:n])
 		if err != nil {
 			return err
 		}
@@ -100,10 +92,10 @@ func New(ws *websocket.Conn) *Conn {
 	// https://godoc.org/github.com/gorilla/websocket#hdr-Concurrency
 	// "Connections support one concurrent reader and one concurrent writer.
 	// Applications are responsible for ensuring that no more than one
-	// goroutine calls the write methods (NextWriter, etc.) concurrently and
-	// that no more than one goroutine calls the read methods (NextReader,
-	// etc.) concurrently. The Close and WriteControl methods can be called
-	// concurrently with all other methods."
+	// goroutine calls the write methods (WriteMessage, etc.) concurrently
+	// and that no more than one goroutine calls the read methods
+	// (NextReader, etc.) concurrently. The Close and WriteControl methods
+	// can be called concurrently with all other methods."
 	pr1, pw1 := io.Pipe()
 	go func() {
 		pw1.CloseWithError(closeErrorToEOF(readLoop(pw1, ws)))
