@@ -146,7 +146,8 @@ func TestWrite(t *testing.T) {
 
 // Test that multiple goroutines may call Read on a Conn simultaneously. Run
 // this with
-//   go test -race
+//
+//	go test -race
 func TestConcurrentRead(t *testing.T) {
 	s, c, err := connPair()
 	if err != nil {
@@ -189,7 +190,8 @@ func TestConcurrentRead(t *testing.T) {
 
 // Test that multiple goroutines may call Write on a Conn simultaneously. Run
 // this with
-//   go test -race
+//
+//	go test -race
 func TestConcurrentWrite(t *testing.T) {
 	s, c, err := connPair()
 	if err != nil {
@@ -259,5 +261,113 @@ func TestClose(t *testing.T) {
 	// internally.
 	if err != io.ErrClosedPipe {
 		t.Fatalf("Write after Close returned %v, expected %v", err, io.ErrClosedPipe)
+	}
+}
+
+// Benchmark creating a server websocket.Conn (without the websocketconn.Conn
+// wrapper) for different read/write buffer sizes.
+func BenchmarkUpgradeBufferSize(b *testing.B) {
+	// Buffer size of 0 would mean the default of 4096:
+	// https://github.com/gorilla/websocket/blob/v1.5.0/conn.go#L37
+	// But a size of zero also has the effect of causing reuse of the HTTP
+	// server's buffers. So we test 4096 separately from 0.
+	// https://github.com/gorilla/websocket/blob/v1.5.0/server.go#L32
+	for _, bufSize := range []int{0, 128, 1024, 2048, 4096, 8192} {
+		upgrader := websocket.Upgrader{
+			CheckOrigin:     func(*http.Request) bool { return true },
+			ReadBufferSize:  bufSize,
+			WriteBufferSize: bufSize,
+		}
+		b.Run(fmt.Sprintf("%d", bufSize), func(b *testing.B) {
+			// Start up a web server to receive the request.
+			ln, err := net.Listen("tcp", "127.0.0.1:0")
+			if err != nil {
+				b.Fatal(err)
+			}
+			defer ln.Close()
+			wsCh := make(chan *websocket.Conn)
+			errCh := make(chan error)
+			server := http.Server{
+				Handler: http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
+					ws, err := upgrader.Upgrade(rw, req, nil)
+					if err != nil {
+						errCh <- err
+						return
+					}
+					wsCh <- ws
+				}),
+			}
+			defer server.Close()
+			go func() {
+				err := server.Serve(ln)
+				if err != nil && err != http.ErrServerClosed {
+					errCh <- err
+				}
+			}()
+
+			// Make a request to the web server.
+			dialer := &websocket.Dialer{
+				ReadBufferSize:  bufSize,
+				WriteBufferSize: bufSize,
+			}
+			urlStr := (&url.URL{Scheme: "ws", Host: ln.Addr().String()}).String()
+
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				ws, _, err := dialer.Dial(urlStr, nil)
+				if err != nil {
+					b.Fatal(err)
+				}
+				ws.Close()
+
+				select {
+				case <-wsCh:
+				case err := <-errCh:
+					b.Fatal(err)
+				}
+			}
+			b.StopTimer()
+		})
+	}
+}
+
+// Benchmark read/write in the client←server and server←client directions, with
+// messages of different sizes. Run with -benchmem to see memory allocations.
+func BenchmarkReadWrite(b *testing.B) {
+	trial := func(b *testing.B, readConn, writeConn *Conn, msgSize int) {
+		go func() {
+			io.Copy(ioutil.Discard, readConn)
+		}()
+		data := make([]byte, msgSize)
+		b.ResetTimer()
+		for i := 0; i < b.N; i++ {
+			n, err := writeConn.Write(data[:])
+			b.SetBytes(int64(n))
+			if err != nil {
+				b.Fatal(err)
+			}
+		}
+	}
+	for _, msgSize := range []int{150, 3000} {
+		s, c, err := connPair()
+		if err != nil {
+			b.Fatal(err)
+		}
+
+		b.Run(fmt.Sprintf("c←s %d", msgSize), func(b *testing.B) {
+			trial(b, c, s, msgSize)
+		})
+		b.Run(fmt.Sprintf("s←c %d", msgSize), func(b *testing.B) {
+			trial(b, s, c, msgSize)
+		})
+
+		err = s.Close()
+		if err != nil {
+			b.Fatal(err)
+		}
+		err = c.Close()
+		if err != nil {
+			b.Fatal(err)
+		}
 	}
 }
